+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning based on [Semantic Versioning](https://semver.org/).
 
 ### Added
 - Added `py.typed` marker for type checkers.
+- Added `qpe::getEnergyFromTrotterPhase` and `qpe::getEnergyFromQubitisationPhase` functions for converting a QPE phase to an energy.
 
 ### Changed
 - Renamed `applyMultiStateControlledPhaseShift` to `applyMultiQubitStatePhaseShift` and `applyMultiStateControlledQubitPhaseFlip` to `applyMultiQubitStatePhaseFlip`.